fixing env and restoring game as an app
