+++ packages/game/nodemon-dev.json
@@ -2,5 +2,5 @@
   "watch": ["src"],
   "ext": ".ts,.js",
   "ignore": ["src/asset-manifest.ts"],
-  "exec": "yarn build:ts"
+  "exec": "yarn start"
 }
