+++ packages/game/src/index.ts
@@ -1,3 +1,6 @@
+import { prepareEnv } from '@adventure-bot/game/env'
+prepareEnv()
+
 import { Client } from 'discord.js'
 
 import {
@@ -10,21 +13,15 @@ let client: Client | undefined
 
 export const getClient: () => Client | undefined = () => client
 
-export const startGameService: (gameOptions: {
-  clientId: string
-  token: string
-  channelId: string
-}) => void = async ({ clientId, token, channelId }) => {
-  createClient({
-    type: 'discord',
-    token,
-    clientId,
-    channelId,
-    onError: (e) => console.error('Discord client error!', e),
+createClient({
+  type: 'discord',
+  clientId: String(process.env.CLIENT_ID),
+  channelId: String(process.env.GUILD_ID),
+  token: String(process.env.token),
+  onError: (e) => console.error('Discord client error!', e),
 
-    onReady: (client) => {
-      gameClock()
-      waitForWinner(client)
-    },
-  }).then((c) => (client = c))
-}
+  onReady: (client) => {
+    gameClock()
+    waitForWinner(client)
+  },
+}).then((c) => (client = c))
